tapmap_db/requirements.txt
psycopg2-binary>=2.9.9
geohash2>=1.1
python-dotenv>=1.0.0

